+++ controllers/user_controller.js
@@ -5,10 +5,17 @@ const jwt = require("jsonwebtoken");
 class UserController {
   static async getAll(req, res, next) {
     try {
-      const user = await users.findAll({
+      const data = await users.findAll({
         include: [addresses],
       });
-      res.status(200).json(user);
+      if(!data){
+        throw { name: "notFound" };
+      }
+      res.status(200).json({
+        status: "success",
+        message: "data berhasil ditemukan",
+        data
+      });
     } catch (error) {
       next(error);
     }
@@ -18,12 +25,16 @@ class UserController {
     try {
       const { id } = req.params;
 
-      const user = await users.findByPk(id, {
+      const data = await users.findByPk(id, {
         include: [addresses],
       });
-      if (!user) throw { name: "Item not found" };
+      if (!data) throw { name: "notFound" };
 
-      res.status(200).json(user);
+      res.status(200).json({
+        status: "success",
+        message: "data berhasil ditemukan",
+        data
+      });
     } catch (err) {
       next(err);
     }
@@ -32,7 +43,9 @@ class UserController {
   static async register(req, res, next) {
     try {
       const { email, name, password, phone_number } = req.body;
-
+      if(!email || !name || !password || !phone_number){
+        throw { name: "nullParameter" };
+      }
       const hashedPassword = await bcrypt.hash(password, 10);
       const newUser = await users.create({
         email,
@@ -44,7 +57,11 @@ class UserController {
         photo_url: null,
       });
 
-      res.status(201).json(newUser);
+      res.status(201).json({
+        status: "success",
+        message: "register berhasil",
+        data : newUser
+      });
     } catch (error) {
       next(error);
     }
@@ -55,18 +72,21 @@ class UserController {
       const { email, password } = req.body;
       const user = await users.findOne({ where: { email } });
       if (!user) {
-        throw { name: "InvalidCredential" };
+        throw { name: "invalidCaredential" };
       }
       const passwordMatch = await bcrypt.compare(password, user.password);
       if (!passwordMatch) {
-        throw { name: "InvalidCredential" };
+        throw { name: "invalidCaredential" };
       }
       const token = jwt.sign({ id: user.id }, "codehorizon");
-      // res.status(200).json({ token });
       res
         .cookie("access_token", token, { http_only: true })
         .status(200)
-        .json(user);
+        .json({
+          status: "success",
+          message: "login berhasil",
+          data : user
+        });
     } catch (err) {
       next(err);
     }
@@ -75,10 +95,9 @@ class UserController {
   static async update(req, res, next) {
     try {
       const { id } = req.params;
-      const { name, email, username, password, phone_number, photo_url } =
-        req.body;
+      const { name, email, username, password, phone_number, photo_url } = req.body;
       const hashedPassword = await bcrypt.hash(password, 10);
-      const updateUser = await users.update(
+      const  [updateCount, [updatedItem]]  = await users.update(
         {
           name,
           email,
@@ -87,12 +106,16 @@ class UserController {
           phone_number,
           photo_url,
         },
-        { where: { id } }
+        { where: { id }, returning: true }
       );
-      if (updateUser == "1") {
-        return res.status(200).json({ message: "User updated successfully" });
-      }
-      return res.status(200).json({ message: "User updated failed" });
+      const message = updateCount === 1 ? "Data berhasil diupdate" : "Data gagal diupdate";
+      const status = updateCount === 1 ? "success" : "error";
+      const data = updateCount === 1 ? updatedItem : null;
+      res.status(200).json({
+        status,
+        message,
+        data
+      });
     } catch (error) {
       next(error);
     }
@@ -101,26 +124,16 @@ class UserController {
   static async delete(req, res, next) {
     try {
       const { id } = req.params;
-
-      // Hapus pengguna berdasarkan ID
+      const data = users.findByPk(id)
+      if(!data){
+        throw { name: "notFound" };
+      }
       await users.destroy({where: { id }});
-      res.status(200).json({ message: "User deleted successfully" });
-
-      // Jika pengguna berhasil dihapus
-      // if (deletedUserCount > 0) {
-      //   // Hapus alamat yang terkait dengan pengguna yang dihapus
-      //   await addresses.destroy({
-      //     where: { user_id: id },
-      //   });
-
-      //   res
-      //     .status(200)
-      //     .json({
-      //       message: "User and associated addresses deleted successfully.",
-      //     });
-      // } else {
-      //   throw { name: "Item not found" };
-      // }
+      res.status(200).json({
+        status: "success",
+        message: "data berhasil dihapus",
+        data
+      });
     } catch (err) {
       next(err);
     }
